+++ packages/cms-users/CHANGELOG.md
@@ -1,5 +1,13 @@
 # zemble-plugin-cms-users
 
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @zemble/mongodb@0.0.8
+  - zemble-plugin-cms@0.0.23
+
 ## 0.0.22
 
 ### Patch Changes